refactor: minor updates after review
