+++ test/casedb/integration/build_db/conftest.py
@@ -1,4 +1,4 @@
-from test.test_client.hooks import rewrite_parametrized_dependency_markers
+from test.test_client.pytest_hooks import rewrite_parametrized_dependency_markers
 
 
 def pytest_collection_modifyitems(items):
