+++ test/casedb/integration/build_db/test_casedb_build.py
@@ -1,6 +1,5 @@
 import copy
 import logging
-from dataclasses import dataclass
 from test.casedb.casedb_test_client import CasedbTestClient as Env
 from test.casedb.integration.build_db.base import TEST_TYPE, VERBOSE  # REPOSITORY_TYPE,
 
@@ -9,6 +8,7 @@ from test.casedb.integration.build_db.create import TestCreate as ModuleTestCrea
 from test.casedb.integration.build_db.delete import TestDelete as ModuleTestDelete
 from test.casedb.integration.build_db.read import TestRead as ModuleTestRead
 from test.casedb.integration.build_db.update import TestUpdate as ModuleTestUpdate
+from test.test_client.pytest_params import BuildDbParams
 
 import pytest
 
@@ -35,17 +35,6 @@ CASEDB_APP_CFGS = get_app_cfgs(
 )
 
 
-@dataclass
-class BuildDbParams:
-    skip_endpoints: bool
-    dev_repository_config: commondb_enum.DevRepositoryConfig
-
-    @property
-    def id(self) -> str:
-        prefix = "skip_endpoints" if self.skip_endpoints else "with_endpoints"
-        return f"{prefix}__{self.dev_repository_config.value}"
-
-
 # Two parameter combinations drive the full Create→Read→Update→Delete chain
 # independently. Each results in a separate isolated database environment (Env):
 #   1. skip_endpoints__SA_SQLITE_EMPTY  — SQLAlchemy/SQLite, bypassing HTTP endpoints
@@ -70,10 +59,6 @@ _PARAMS = [
     ),
     BuildDbParams(
         skip_endpoints=True,
-        dev_repository_config=commondb_enum.DevRepositoryConfig.DICT_EMPTY,
-    ),
-    BuildDbParams(
-        skip_endpoints=False,
         dev_repository_config=commondb_enum.DevRepositoryConfig.DICT_EMPTY,
     ),
 ]
